pyproject.toml
[project]
name = "cogheat"
version = "0.1.0"
description = ""
authors = [
    {name = "Guillaume Beraudo",email = "guillaume.beraudo@camptocamp.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "shapely (>=2.1.1,<3.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
